voice_survey_android_app/voicebotserver/requirements.txt
backports.functools-lru-cache==1.5
certifi==2019.9.11
chardet==3.0.4
configparser==4.0.2
dj-database-url==0.5.0
Django==2.2.5
django-extensions==2.2.5
django-heroku==0.3.1
django-mysql==3.2.0
djangorestframework==3.10.3
googletrans==2.4.0
gunicorn==20.0.4
idna==2.8
indic-transliteration==1.8.9
joblib==0.14.0
Morfessor==2.0.6
nameparser==1.0.4
numpy==1.17.2
pandas==0.25.1
polyglot==16.7.4
psycopg2==2.8.4
psycopg2-binary==2.8.4
pycld2==0.31
python-dateutil==2.8.0
pytz==2019.2
regex==2019.8.19
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.3
selenium==3.141.0
six==1.12.0
sklearn==0.0
splinter==0.11.0
sqlparse==0.3.0
textdistance==4.1.4
urllib3==1.25.6
virtualenv==16.7.5
whitenoise==5.0.1
xlrd==1.2.0
